Make all mutation operations dec-ref their original value on COW
Summary:
All common array mutation operations (Append, SetInt, SetStr) are now move operations. Compared to the non-move variants, these operations do not inc-ref the value being inserted and are responsible for dec-refing the old array if a new array is created.

In doing so, we also normalize the SetNewElem{Vec,Dict,Keyset} opcodes to all be move operations on the value.

A forthcoming diff will do the same for the cold mutation operations (Remove{Int,Str} and Pop).

Reviewed By: ricklavoie

Differential Revision: D24682826

fbshipit-source-id: 069b3aab3b0230b88872a71021b8c5ffc527d5d7