Crash less
Summary: When we get an unexpected kind of node, we often panic in the direct decl parser. All of these panics should be unreachable when parsing a file with no syntax errors, but some may be reachable in files with parse errors. Even in the presence of a parse error, some of these panics may be unreachable in practice--the parser might never pass a node of an unexpected kind. Unfortunately, there's no easy way for us to guarantee that these panics are unreachable (and even if some are unreachable now, they may become reachable after future changes to the FFP). In the interest of avoiding crashing on files containing syntax errors, this change has the direct decl parser silently ignore unexpected node kinds instead of raising a panic.

Reviewed By: Wilfred

Differential Revision: D26092811

fbshipit-source-id: a299e8e9c66c96dbfb0ab4d5d45795cdb19f17dd