Delete some expression parsing
Summary:
Parsing these kinds of expressions was necessary when we stored expressions in class constants, but we don't do so any longer. We can remove these smart constructors with no ill effects.

We still need some expression smart constructors because of the `infer_const` behavior (where type hints can be omitted for class consts with a simple initializer). If we require type hints on all constants, we may be able to remove the remaining expression parsing logic.

Reviewed By: Wilfred

Differential Revision: D26092165

fbshipit-source-id: 2ec74735b8dccefc245306e5b733648fb38ab452