Add parsing support under UnstableFeatures for readonly parameters and return
Summary:
This diff adds very basic parser support for readonly parameters and returns.

Every function can take in readonly parameters and its return type can be marked readonly. The added test case illustrates all of the parsing cases necessary, including:
- readonly expressions
- readonly on properties
- readonly as a function modifier

^ the above are not yet implemented, and will be as another diff on top.

So far, we parse these and ignore them both in HackC and in the typechecker, and only if EnableUnstableFeatures("readonly") is on.

I considered calling this thing "param_mutability" and "function_mutability_return" instead of readonly_return to be more generic, but unfortunately, those already refer to things that exist in the Reactivity world. Once we get rid of those with coeffects, we can revisit changing the implemented name to mutability to capture more than just readonly here.

In an effort to make the diff reviewable, I put only the actual source changes into a different diff(D26055162), plus the one extra test I added that tests the feature itself.

After this diff was accepted, I merged them into a single diff here.

Reviewed By: zilberstein

Differential Revision: D25909167

fbshipit-source-id: 0d7e2914a8779e23f07caffb97308be3e58db031