Do not mark ArrayData::release as noexcept
Summary:
Marking functions as noexcept prevents tail call optimization, as the standard
guarantees that if a child function throws, then the noexcept function will
invoke std::terminate. This is only possible if the frame of the noexcept
function is present when unwinding.

Removing the noexcept specifier allows ArrayData::release to perform a
tail-call.

Differential Revision: D26083031

fbshipit-source-id: 3c717b31a1ec7239b3c5b1306b79dda8c8c50e71