Better error messages for non-object access
Summary: The non_object_member error would previously refer to a type constant as a property because of the binary ~is_method argument. There's already a polymorphic variant `class_member_kind` used for static members that is used elsewhere in Errors, so I extended it to cover methods and properties and used that.

Reviewed By: losvald

Differential Revision: D26120916

fbshipit-source-id: 31b8d254f38b21456ab1fdef494c549357388a53