check variance of type definition using AST instead of decl
Summary:
A later diff will do the same for class definitions.

This is in preparation to enforcing primary error positions from AST.
The class def counterpart of this diff will also have the advantage of checking the only current class members and not its inherited members, which should perform marginally better.

Reviewed By: andrewjkennedy

Differential Revision: D26126763

fbshipit-source-id: 2ff2b874f9daa3e2de47f3f04e94ee705130cd01