Interpret varray as vec and darray as dict if flag is set
Summary:
Implements varray->vec and darray->dict behavior for AST decl and the direct decl parser.

In the AST, the Naming module is responsible for converting `Happly "varray"` to `Hvarray` (ditto for primitives).

For the direct decl parser, the `make_varray_type_specifier` smart constructor directly builds a Tvarray. Due to some ambiguities in the grammar, the DDP's path for bare `varray` is slightly different, but it is covered too.

Reviewed By: dneiter

Differential Revision: D26012191

fbshipit-source-id: f0a32d15f4da47b9572a60077c1280f4194da3ee