Drop duplicate meth_callers from repo build
Summary:
In the runtime we just ignore duplicate meth_caller functions when loading units. This used to work fine in repo mode but when we implemented repo autoloading we banned duplicate functions which caused builds with identical meth_callers in different units to fail.

Since unit loading in repo mode is unobservable just deduplicate meth_caller implementations and keep the first one we see (they're already included in the repo autoload map).

Reviewed By: ricklavoie

Differential Revision: D26208508

fbshipit-source-id: 0cfd177521b7020195b64177396dbe6eb63844f7