Fix indentation of var_dump for class_meth
Summary:
The code for dumping ClassMeth is hard-coded to dump no leading indentation, and then four spaces of internal indentation.

The diff corrects this so that ClassMeths are dumped with proper indentation (inheriting leading indentation and properly respecting 2 or 4 spaces of indentation depending on `var_dump` or `print_r`, respectively.

Reviewed By: ricklavoie

Differential Revision: D26135885

fbshipit-source-id: f5a8165eca4e8c92cf26c1df09416ad52c401446