Virtualize void returns
Summary:
It has been requested that void lambdas within Expression Trees (lambdas that do not return a value) have void "virtualized" with a client void type.

This is accomplished as follows:

- If a lambda has no `return expression;`, we do not make any changes, (IMPORTANT: Even if that expression is `void`)
- Else: If a lambda's last statement is not a `return;` or `return e;` we append a `return;` statement to the end of its body.
- Finally: All `return;` statements in the Expression Tree context are virtualized as `return ${ Visitor::voidLiteral() }`

Thus, the following virtualization occurs:

```
Visitor`
  () ==> {}
`;
```

Is virtualized as:

```
`() ==> { return ${ Visitor::voidLiteral() }; }`;
```

Another example where we virtualize void and add the void return at the end
```
Visitor`
  () ==> {
    if (true) {
      return;
    }
  }
`
```

is virtualized as

```
`
  () ==> {
    if (true) {
      return ${ Visitor::voidLiteral() };
    }
    return ${ Visitor::voidLiteral() };
  }
`
```

Another example. We do not do any sophisticated flow analysis, so if a body already terminates, we actually will go ahead and add another return. For instance:

```
Visitor`
  () ==> {
    return;
    1;
  }
`
```
is virtualized as

```
`
  () ==> {
    return ${ Visitor::voidLiteral() };
    ${ Visitor::intLiteral(1) };
    return ${ Visitor::voidLiteral() };
  }
`
```

Similarly:

```
Visitor`
  () ==> {
    if (true) {
      return;
    } else {
      return;
    }
  }
`
```

is virtualized as:

```
`
  () ==> {
    if (true) {
      return ${ Visitor::voidLiteral() };
    } else {
      return ${ Visitor::voidLiteral() };
    }
  return ${ Visitor::voidLiteral() };
  }
`
```

The cost, unfortunately, is that we will do some extra computation for the spliced expressions (since those are lifted and processed before the function body).

Reviewed By: Wilfred

Differential Revision: D26008099

fbshipit-source-id: 5f880c1edb9063755378960fdaa98014708c8a88