Error on unbound variables from outside scope
Summary:
Currently, when we desugar an Expression Tree, we desugar to Hack style lambdas. Hack style lambdas have the usually nice side effect of implicitly capturing any variables used in them from the outer scope. However, in the case of Expression Trees, we do not want this behavior. Consider:

```
function test(): void {
  $x = new VisitorInt();
  Visitor`$x`;
}
```

Any variables that are going to be introduced from the outer context need to be spliced into the Expression Tree. As of right now, we do not produce an error here because the typechecked lambda captures `$x` from the outer scope.

One fix to this problem is, instead of desugaring to Hack style implicit variable capture lambdas, is to desugar to PHP style lambdas which require explicit variable capture. Then we only capture spliced in variables for correctness.

Thus, simplified, instead of desugaring for typechecking the example above into
```
function test(): void {
  $x = new VisitorInt();
  (Visitor $_) ==> { $x; };
}
```
We might desugar this into:
```
function test(): void {
  $x = new VisitorInt();
  (function(Visitor $_) use () { $x; };
}
```
The former would not throw an error. The latter would, which is what we want.

Reviewed By: Wilfred

Differential Revision: D26112278

fbshipit-source-id: 05407c8d01e8dac4df8049a2ca0c2b1de78ae1b7