Refactor/optimise expected & declared lambda parameter consolidation
Summary:
To select between declared and expected types, we were passing lambda parameters around which derive the declared type. This diff elides the lambda parameters and instead uses declared parameter type to see if the hint was present.

This diff also makes the recursive function over parameters terminate earlier in the base case.

Reviewed By: andrewjkennedy

Differential Revision: D26149032

fbshipit-source-id: a61102933723e9d4bd436785796029ac53718b76