Visit implemented interfaces with their traits
Summary:
- Logical precursor to constants in traits

Currently, when we look for constants on interfaces, we crawl over `m_interfaces`, which contains:
- interfaces from parent
- interfaces from self
- m_interfaces of self's interfaces
- interfaces from used traits
- interfaces Stringish and XHPChild (when required)

This appears to be duplicate work, as our parent class should already have resolved the constants from our parent's interfaces, and our interfaces should have already resolved the constants from their interfaces. To better reason about this code / not perform duplicate work, let's tease out which interfaces we actually might pull constants from (declInterfaces + interfaces from traits).

In order to get the precedence rules we want for constants in traits, we also want to look at these interfaces when we look at the trait that implements them. This change is really only necessary in non-repo mode, as hhbbc already visits them like this, and can therefore already resolve which one to use.

RFC:
Currently for hhbbc, we don't flatten "implements" clauses from traits onto the classes. We instead just visit these in addition to declared interfaces in the runtime from a list of "interfaces from used traits" we generate. See [here](https://fburl.com/diffusion/u3gu0yc9). We could flatten these clauses. We would need a few checks for sealed classes, and some bitfield to keep track of which interfaces are "from traits" to know whether to fatal on conflict. It's unclear to me if this is better.

If we did this, we would't need to specially visit the implemented interfaces in the runtime, as we'd just visit them with that class's declInterfaces.

Reviewed By: ricklavoie

Differential Revision: D26141022

fbshipit-source-id: c64f0f7a55f093f24cb0c906b40a6e5013fa3611