Eliminate HAM specialization intrinsics
Summary:
Eliminate the dummy_cast_to_kindofarray, ...darray, and ...varray intrinsics.

The ...array and ...darray ones are equivalent to darray casts, while the ...varray one is equivalent to a varray cast. But only the ...array one is used in tests, which I've replaced with the cast. Per bnh's suggestion, I've also simplified darray(varray[]) and darray(vec[]) to darray[].

Differential Revision: D26304046

fbshipit-source-id: b13e6728212bfe12aaafca725baa53cfef9a6480