Add autocompletion for XHP enum value with constants
Summary:
The goal of this diff is to add autocompletion support for enumeration values of XHP attributes with enum types. This covers one of the scenarios outlined in https://fb.quip.com/fDxWA9VMQo8i

Specifically, if we are autocompleting an XHP attribute whose type is an enum class with constants:

```
enum MyEnum: string as string {
  TYPE_A = "A value";
  TYPE_B = "B value";
  TYPE_C = "C value";
}

class :foo:bar {
  attribute MyEnum my-attribute;
}
```

then if we type

```
<foo:bar my-attribute={
```
Autocompletion should show us

```
MyEnum::TYPE_A
MyEnum::TYPE_B
MyEnum::TYPE_C
```

This diff adds logic to add the enum value constants to the list of autocompletion suggestions for this specific scenario.

Reviewed By: Wilfred

Differential Revision: D26187287

fbshipit-source-id: dc091ef7b394a2bff5728b5b2b36c4329698bc27