Don't use non-parental braces around cast expressions
Summary:
hackfmt supports not breaking inside the parentheses of a function call in some circumstances:

```
// We produce:

f(vec[
  $foo, // single-line comment
  $bar,
]);

// Instead of:

f(
  vec[
    $foo, // single-line comment
    $bar,
  ],
);

// And we produce:

map($vec, $element ==> {
  // ...
});

// Instead of:

map(
  $vec,
  $element ==> {
    // ...
  },
);
```

But this isn't always appropriate--for some constructs, this formatting looks pretty weird:

```
f($x ==>
  do_something_with($x));

f($x
  ? $y
  : $z);
```

So we blacklist them from this special behavior.

This behavior is currently kicking in for `is` and `as` expressions, producing results like this:

```
f('%s input is not valid', $y__________________________
  as string);
```

This doesn't look so great. It's easy to disable, though--we just need to add these expression types to the blacklist.

Reviewed By: dakotareier

Differential Revision: D26290833

fbshipit-source-id: d3ff60e9674c1dc5d014f231adc37a2dbafae77a