Handle unavailable functions in emitResolveMethCaller
Summary:
In non-repo mode while the meth_caller is guaranteed to be defined in the same unit as the `ResolveMethCaller` bytecode it may interfere with an identical meth_caller elsewhere and be deduplicated. I have some ideas for handling this differently to avoid the interpOne if we need to but this was the easiest fix for now.

In repo mode we do the deduplication at the time the index is built. While we could load the unit inside the JIT (now that unit loading is unobservable in repo mode) this was a straightforward fix which between jumpstarting and profiling translations is unlikely to affect hot optimized translations.

Reviewed By: jano

Differential Revision: D26280489

fbshipit-source-id: 7ac6faf2a49c79df795881c514080edd5cf7b61f