Address fb_compact_serialize HAM behaviors
Summary:
fb_compact_serialize does some funny logic on PHP arrays. If a darray is "nearly list-like" - that is, if it is non-empty and its keys are all ints in [0, 2 * size] - we'll encode it as a list with skipped fields. Likewise, if a varray is empty, we'll encode it as a map (!)

In order to make this code HAM-safe, we make a couple changes to its behavior:
1. Post-HADVAs, we run marked vecs through the empty logic and marked dicts through the "nearly list-like" logic.
1. To let us fix sinks before flipping the big switch, we add an option to apply this logic to all Hack arrays.

We expect the new option to be usable in most places we call fb_compact_serialize, because its deserializer only produces darrays.

Reviewed By: dneiter

Differential Revision: D26321253

fbshipit-source-id: e8fe6b3be511207422f69bd5a0042f99017dbd0f