Require write_props for accessing an enum class variant
Summary:
When accessing a variant on an enum class, the behavior is that you lazily evaluate the `new ClsName(args)` in your current context, and enum classes require write_props.

So, when we access a class constant and it's an enum class, we add a check to make sure the local capability includes WriteProperty.

Reviewed By: losvald

Differential Revision: D26353430

fbshipit-source-id: de24a1096d8c22c84ebc61db382f939efa30726f