Change the runtime inferred lambda to null
Summary:
Currently, for the inferred type trick, we pass a closure of `() ==> { throw new Exception(); }` to the runtime factory method.

Constructing a closure can be an additional expense that we don't want to deal with. One way to reduce this cost is to use something cheaper to construct, like `null`.

This diff changes the runtime value to `null`. In addition, we change the type of the `makeTree` method to take in a nullable function to make the runtime check work out.

Reviewed By: Wilfred

Differential Revision: D26357985

fbshipit-source-id: ff2402fa87b71a454bd448dc8b1e9fbd20892cd5