Back out "Add readonly info to calls and obj_get"
Summary:
Original commit changeset: a58509f1dab0

Implementing readonly checks at the AST level was our original design, but after writing a lot of code I've decided that we're adding way too many nodes to the AST for the sole benefit of sharing a single piece of code. Thus, I think it would be better to implement this as a TAST check, and duplicate the local checks in the emitter, rather than trying to share the code and polluting so many AST nodes.

Reviewed By: zilberstein

Differential Revision: D26343326

fbshipit-source-id: c3d5c1f407afe3bf355cc19e265441382f1e64ee