add flag for stricter string concatenation and interpolation
Summary:
Added the flag "--enable-strict-string-concat-interp". When this flag is set, string interpolation and concatenation operations will require arguments to be type `arraykey`, though this stricter check is implemented in the next diffs in the stack.

This is part of the steps described in [this doc](https://fb.quip.com/3yECAnOtOPT0).

Reviewed By: vassilmladenov

Differential Revision: D26353862

fbshipit-source-id: c89d63076e224e25fa00b0d724561a382ee49e49