Implement readonly as a tast check
Summary:
This diff begins to implement readonly as a standard tast check. It implements basic readonly checks like RO_MODIFY and RO_PARAM, RO_DEEP, RO_THIS, VARIABLE_CONSISTENT_RO from the doc:
https://fb.quip.com/UFsrA3IdxWLY

Implementation detail:

Unfortunately, there isn't a clean, immutable way in our code to do a "tast check with custom env". The "tast_visitor_iter_with_state" that we use has a state, but the state isn't updated between runs.

The cleanest way I can figure out how to store a context without having to reimplement the entire visitor is to have a mutable context which is cleared at the start of each function call. Then, on each assignment, we add the readonliness of each variable to the context.

When we later want to implement scoping for readonly, we'll have to merge contexts together at join points; but for now, this is reasonable.

Reviewed By: zilberstein

Differential Revision: D26347406

fbshipit-source-id: 6e7740b5d56aeeb4d94c782c00cca785a53cf97a