Add Readonly immediate to bytecodes 1/2
Summary:
Add ReadonlyOp immediate to bytecodes:
```
- {SetOpM}
- {InitProp}
- {CGetS, SetS, SetOpS, IncDecS}
- {BaseSC}
```
For the member base operation bytecodes, we will store the `ReadonlyOp` in the MemberKey because otherwise, we would need to increase the size of the Bytecode to 40 bytes for lowptr builds.

- ReadonlyOp can be `{Readonly, Mutable, Any}`
- Emit dummy value `Any` for now from HackC.

Note: Defined READONLY_OPS in member-key.h because hhbc.cpp includes member-key.h, and we need it defined in member-key.cpp as well for diff 2/2.

Reviewed By: oulgen

Differential Revision: D26335799

fbshipit-source-id: 151f12632df1989dcc332e43c300861b07bb298f