prevent enum inclusion to reference classes that have not yet been loaded by HHVM
Summary: HHVM was not handling correctly enum inclusions dependencies when loading classes.  A diff is on the way (D26410129), but to prevent a typehole this diff makes Hack return an error if enums are declared in an ordering on which HHVM fatals.  This diff will be reverted when D26410129 reaches the HHVM release.

Reviewed By: madgen

Differential Revision: D26412090

fbshipit-source-id: 018c7db30cfac6c6abc18f0ec8480afb48394255