track enum inclusion dependencies when loading classes
Summary:
The first user of enum inclusion discovered a glitch in the HHVM support. The code below:

```
enum B: string as string {
   use A;
}
enum A: string as string {
   FOO = 'foo';
}
<<__EntryPoint>>
function main() : void {
   print B::FOO;
}
```

fails with

```
$ php ~/foo5.php
Fatal error: Undefined enum: A in /home/fzn/foo5.php on line 6
```

This diff updates the logic in `Class::def` to deal correctly with the case above.

Reviewed By: WizKid

Differential Revision: D26410129

fbshipit-source-id: 85a0821ead84737a82410518f7cf174e95c02437