Abstract bit for context constants in traits
Summary:
Preclasses determine whether a constant is abstract by checking whether it has a value or not.

Context constants never have a value.

So, we need a bit to determine whether they are abstract in order to properly handle conflicts.

------------------------
Additional Notes:

There might be a way to just look at the coeffects field, but it looks like StaticCoeffects::none() == [], so I'm not sure we could tell the difference between a "pure" concrete ctx constant and an abstract one.

[RFC] Would it be better space-wise to instead create a dummy-value for tvaux and store the `isAbstract` information in its `constModifiers`? Could imagine creating something in as.cpp to pass around.

Reviewed By: oulgen

Differential Revision: D26414769

fbshipit-source-id: bbe66e80494d445834fdc99799f9f7844bfc3c41