Speed up CLI server init times
Summary: We're re-compiling systemlib files and installing extensions on every HHVM process, even if that process is just talking to the CLI server. Create a special init path for CLI server mode that skips a lot of the slower pieces so we don't have to do them if we successfully talk to the server.

Reviewed By: paulbiss

Differential Revision: D26187757

fbshipit-source-id: 857517cbe45b8d1a29aa191f4e491c3574a878da