Fix context constants codegen in hackc
Summary:
This diff fixes the bugs in abstract/concrete context constant codegen in hackc.

Given a Hack source code as
```
abstract class A {
  const ctx C1 = [];
  abstract const ctx C2;
  abstract const ctx C3 = [];
  const ctx C4 = [random_coeffect_who_am_i];
}
```
We need to emit the coeffects of concrete context constants of if all the contexts are non runtime enforced, then emit `defaults`.
If the abstract context constant has a default value, we should emit it as a concrete context constant. If it doesn't have a default value, emit it without any values which will indicate to HHVM that this context constant is abstract.

```
.class {} [abstract] A (3,8) {
  .ctx C1 pure;
  .ctx C2;
  .ctx C3 pure;
  .ctx C4 defaults;
}
```

Reviewed By: kavoor

Differential Revision: D26434854

fbshipit-source-id: 0d60828a264be7936faa94d36a236197a0ce50e1