Fix return type inference for in-out builtins
Summary:
The logic to calculate the return types for builtins was
ignoring in-out entirely. So if a builtin returned bool, with an int
in-out param, we'd report the return type as simply bool. The only
reason this wasn't causing problems was that the code to unpack the
return type just returned TInitCell for anything that wasn't a
specialized vec. Fix this by calculating the return tuple correctly,
and add an assertion that the return value could at least be some kind
of vec.

Reviewed By: paulbiss

Differential Revision: D26431215

fbshipit-source-id: e57c11d61be3e1863ed02c980379f0804aed57ca