Make Hack tests using XHP literals safe
Summary:
XHP literals in Hack such as `<p>Hi</p>` are desugared into constructor calls. In www, a handful of files define these constructors for nearly all uses of XHP, but these are not accessible in our typechecker tests.

In anticipation of D25612305 where we start type checking XHP constructors, we add an HHI file with a class with the expected constructor and update the existing tests to use it.

Reviewed By: ljw1004

Differential Revision: D25681301

fbshipit-source-id: e0661632cc134c7a1aed2e9cbedf0b808a00be5c