Typecheck XHP constructors
Summary:
Currently, XHP constructors are called in two different ways: directly and via desugaring. Directly, if we have a class `:div` with a unary constructor, we can write `new :div(42)`. Indirectly, when we have an `Xml` node `<div colur="blue">Hi</div>`, this gets desugared into a constructor call as well. So far we weren't typechecking XHP constructors at all, meaning completely meaningless programs type check.

This diff fixes it. See the test files for example type errors that we can catch now.

This diff does not eliminate Xml nodes from the AST. We temporarily generate New expressions within the type checker so that they can be type checked properly and then return a typed Xml node as usual.

Because it typechecks the constructors, we now ensure that an appropriate constructor is defined for the XHP class somewhere in its type hierarchy.

Reviewed By: Wilfred

Differential Revision: D25612305

fbshipit-source-id: a4a6856fd841e561528153241c56dd5b37b857f7