Add more complicated unit tests for lambda coeffects
Summary:
The previous diff added a very simple test because warning/enforcement is disabled by default. This diff was supposed to be included but it was yanked out because the tests randomly failed.

It wasn't random - it failed in repo mode and passed otherwise. The two tests that failed were the first two tests of this patch.

It appears that because the function body for `regular` was empty, hhbbc optimized away the function calls. This was the exact issue with the test (the first two expectations were never met).

This hacks in a way to avoid this by happening - by adding a function with a side-effect that cannot be optimized.

Reviewed By: oulgen

Differential Revision: D26384027

fbshipit-source-id: 2e26f46e86c939f6524dffe547f28b73114b5f49