Add better error for meth_caller with magic methods
Summary:
We currently throw an `No instance method` error when one uses a magic method in `meth_caller` (`meth_caller('Foo', '__construct')`).

We can make this error more consistent if we also throw the `magic method` error, such as we do with regular method calls like:

`(new Foo())->__construct()` which also throws:

```
`__construct` is a magic method and cannot be called directly
```

Since we don't treat `__construct` as a method for Typing_obj_get, I've removed the error path in this case, so we only show one error to the user and it's more clear.

Reviewed By: periodic1236

Differential Revision: D26210840

fbshipit-source-id: bc45ab6878c224345666e6bfaac568a027105a61