Use folly::ConcurrentHashMap instead of tbb::concurrent_hash_map
Summary:
Folly version is lock-free on reads.

Compared to the previous version, if the first few accesses to a given spec race with each other and all get to line 80, there will be multiple parallel calls to `compileSpec`, but only one will win (since `try_emplace` only inserts if the key didn't exist) and the rest will be discarded. It would not affect cache correctness or cause any memory leaks, so it should be fine.

Reviewed By: ricklavoie

Differential Revision: D26473375

fbshipit-source-id: e6da5db9faa6a45e2e0b68ade89e52cef8d75d46