Add trivia factory
Summary:
To enable positioned parser with full trivia, we need to support full trivia list in `PositionedToken`. Full trivia list allocates a vector to store all trivium, this list has be allocated in arena for two reasons,
1. `PositionedToken` is allocated in arena, if it contains `std::vec::Vec`, memory leaks. Because releasing memory in arena doesn't call `Drop`.
2. allocating in arena is faster.

To allocate trivium in arena, we need a state full trivia factory, which keeps a ref to arena.

This diff refactors token factory to include a trivia factory.

Reviewed By: shayne-fletcher

Differential Revision: D26296227

fbshipit-source-id: b16ff369b03dd58ed9d1dcb7bb748c17415c9a75