Test for interpolated string
Summary:
Adding a test to make sure that we don't allow string interpolation. We currently allow both single and double quoted strings. Since double quoted strings have variables interpolated, adding a test for it.

Ultimately, interpolated strings resolve to a different node in the AST but just making sure we ban it for the moment.

Reviewed By: Wilfred

Differential Revision: D26498018

fbshipit-source-id: 7b549f03ebc0e43dd48f43f2fafab83570211925