decl type for constants should be const_decl instead of decl_ty
Summary: The decl heap only stores a decl_ty for each global constant, but some other places in the code base use a Typing_defs.const_decl. There is a place where a const_decl is reconstructed from a decl_ty, and other places where a const_decl is stripped down to a decl_ty. With this diff, we use const_decl everywhere.

Differential Revision: D26366926

fbshipit-source-id: 337f5f7473f1127a8528602d6fe2ab40b4fd961c