Basic property accesses
Summary:
This diff adds basic support for property access checks on readonly types.

It only works right now for the basic case, when the type of the receiver of a property access is a single class.

When the class is a type variable, a generic, or union/intersection, we'll need a bit more complicated checks. I added a TODO for that for now

Reviewed By: zilberstein

Differential Revision: D26383467

fbshipit-source-id: 6416a4881a7b41342c19e3b7d4735483a3e79319