Fix refcounting issues for marking bespoke arrays
Summary:
The setMove operations here are incorrect. result->incRefCount(); may operated on result after the move has destroyed it.

It's hard to see this breakage, because apply_mutation_slow is only used for bespoke array-likes - all vanilla arrays go through a faster path.

Fix the bug by forcing a copy whenever we have to change one of the bespoke array's elements due to a marking call. This copy doesn't change the algorithmic behavior of this operation, because we're already doing an Theta(n) iteration. In a follow-up, I'll try to remove this copy.

Reviewed By: colavitam

Differential Revision: D26507225

fbshipit-source-id: 41e031b94b578e0f188175c8ab9b049c18002537