Add pure context to `extract_standalone_make_default` generated function
Summary: `hh --extract-standalone` generates a function `extract_standalone_make_default` whenever at least one extracted function parameter has a default value. Although the function is _annotated_ as pure, with `<<__Pure>>`, the type checker will complain when it is called from a pure function as it does not have an explicit pure context (i.e. an empty intersection of contexts). This diff updates the function with an empty contexts and modifies related tests.

Reviewed By: madgen

Differential Revision: D26512234

fbshipit-source-id: b155b1f3de3f64c6b6dd15dcb97df0929a06799e