Clean up use of get_self
Summary: The `get_self_ty` helper produces a `Tany` if we're checking a top-level function. This seems like a bad smell - instead, we should let the caller deal with a `None`, as it does with the other helpers `get_self_id` and `get_self`. Also take the opportunity to use `get_self_id` where possible, instead of deconstructing a class type.

Reviewed By: vsiles

Differential Revision: D26544212

fbshipit-source-id: 1afce61fdf9dff9d53fbb344338da338a0f08a7e