Missed setLocalsDecRefd causes backtrace corruption
Summary:
I believe this bug is responsible for the unwinding/backtracing crashes in c3.

Suppose we are in a wait handle structure that will introduce a cyclic dependency when the current wait handle completes. For example:
- A awaits B
- B awaits AwaitAllWaitHandle(vec[C, A])

When C completes, the AwaitAllWaitHandle will be unblocked and await A, throwing an exception. However, if the RetC of C is interpreted, we will first decref the frame locals of C, then invoke the `ret` handler of its WaitHandle. The `ret` handler will unblock the AwaitAllWaitHandle, which throws an exception. Backtracing will then collect the $this pointer (and potentially metadata pointers) in C if they are available, as it is active on the VM stack.

Since we forgot to set LocalsDecRefd for C in the RetC interpret case, backtracing will pull a dead pointer to C's $this or metadata and produce memory corruption.

Differential Revision: D26583668

fbshipit-source-id: 9391af6788cdca42e9ecd26a28777f8543fc0dee