Make the vanilla "layout index" non-zero
Summary: We're going to use an unordered code to encode bespoke layouts, and test instructions to test them. This approach can't work if the vanilla layout index is 0x0000. Change this index to 0xffff instead, to give us maximum flexibility with bespoke layout indices. Note that vanilla arrays don't need to be tested as an unordered code, so it's fine to choose this maximally-pessimistic value.

Reviewed By: colavitam

Differential Revision: D26462894

fbshipit-source-id: 941b7747caebe057b6851dec93e75ac5558a144a