Do not set expected type if it's an unsolved type variable
Summary: Expression type inference takes an optional parameter `expected` that is the "expected" type of the expression, if known. It makes no sense to pass in an unsolved type variable as the expected type, and is the apparent cause of an issue with type constants (see test). The issue is fixed if we avoid setting the expected type in the case that it's a type variable.

Reviewed By: DavidSnider

Differential Revision: D26132034

fbshipit-source-id: 0bf8f293dd633b320e3c27ff1d09cd3d06a9feb0