Disable teardown opts in debug builds
Summary:
The third (and hopefully last) unwinding crash.

The teardown optimizations are responsible for the crashes in dbgo servicelabs
(see test plan). Disable them for now until the cause is investigated.

oulgen: I believe you're most familiar with this optimization. This crash
appears to arise when a request timeout occurs while in a CPP builtin.
Unfortunately, I haven't been able to create a local repro, but I suspect
something is not quite right with the optimization wrt surprise checks.

Reviewed By: oulgen

Differential Revision: D26609968

fbshipit-source-id: cbf63fdc23d0b49ac8ea18330d38a3e306fc0f9a