Allow for arbitrary type unions in HHBBC
Summary:
Remove the predefined restriction for types in HHBBC. Types can now be
arbitrary unions. We still have predefined types, but they are only
the types with predefined names and have no real specific significance
otherwise.

The types are now generated with macro magic, which makes it easier to
add new ones. One simply has to add to one line to a macro and the
rest is taken care of.

I've taken a pass across all of HHBBC and made sure the logic worked
with arbitrary unions (for the most part it was already correct). For
a few critical areas (interp-minstr.cpp) in particular, I modified the
logic to project across each set of types, then union together the
possible results.

Removing the predefined restriction actually simplified a lot of logic
within type-system.cpp (though some aspects got more complex). For
example, union_of became simpler and more efficient, as it can now be
implemented as a union of bits, along with a quick specialized data
check. intersection_of is similar. A number of places that had to go
out of its way to maintain a predefined type can now be implemented
with bit operations.

As part of the cleanup, I added a lot more invariants, particularily
with regards to arrays. I codified some of the rules about specialized
data more, formally introducing the concept of "support" bits and what
they actually mean. Array related specialized data is one of the
things that become a lot more complex when the predefined restriction
was removed. That's because we can now have types like
{TKeyset|TDictish}[TArrKey:TStr] for example. So we need to enforce
the "upper" and "lower" bounds of what types can be in the
specialization, which depends on the array bits. This makes
intersection_of, in particular, more complicated. Using the above
example, if you intersect away the TDictish, you need to conform the
specialization to obey TKeyset invariants.

All of the elem/set/newelem array functions can now be unified into a
single set of 3 functions. There's no longer any need for different
ones for each array type (and in general would be hard to support
since arrays can be unions of them). Instead these functions will
automatically split the array into its specific types, perform the
operation on each, then union the results together.

The logic for tracking marked legacy arrays and array provenance tags
was a mess, and didn't seem correct. In addition, it caused several
invariant violations. Rework that logic completely. Encapsulate
everything into a class called HAMSandwich, which defines the
appropriate union/intersection/subtype/etc operators. Every Type has a
HAMSandwich (luckily we can fit it into 8 bytes and therefore not
expand Type), so most of the logic can remain blissfully unaware of
its details. It should be easier to remove when the time comes too!
With this rework, we can once again enforce the invariant that empty
arrays do not support specialized data.

I found the existing gtests deficient, so I added a lot of new ones
and modified the existing ones. We now generate most of the test cases
programatically, which revealed quite a number of bugs in the
implementation. Also add a few new unit test which demonstrated more
complicated problems. To make the gtests more effective, I added some
pseudo-private helper functions to generate types more directly
(helpful for pathological cases).

Make the hhbbc stats dump more programatically generated since the old
one was missing a lot of types.

Do some other cleanup while I'm around. Namely, finally replace all
asserts with assertx. Also use the B* types more consistently.

Sorry for the huge diff. It was hard to split this into any smaller
pieces.

Differential Revision: D26566728

fbshipit-source-id: e80b82792c04ffcc71b0582ea1eb154b0429d3c1